表单自动生成列、宽度自适应

+++ src/Blazui.Component/Table/TableHeader.cs
@@ -10,7 +10,7 @@ namespace Blazui.Component.Table
     public class TableHeader<TRow>
     {
         public string Text { get; set; }
-        public int? Width { get; set; }
+        public float? Width { get; set; }
         public string Property { get; set; }
         public Func<TRow, object> Eval { get; set; }
         public bool IsCheckBox { get; set; }
